requirements.txt
python-gnupg
fs
